fix: Windows/Linux CI compatibility
- Move fasthttp tests to build-tagged files
- Fix transport_test.go to not reference fasthttp types on Windows
- Fix config_transport_test.go to be platform-agnostic
- Pin golangci-lint to v1.64 to avoid Go version mismatch
