+++ transport/transport_test.go
@@ -2,8 +2,6 @@ package transport
 
 import (
 	"testing"
-	
-	"github.com/valyala/fasthttp"
 )
 
 func TestHandlerFunc_ServeKruda(t *testing.T) {
@@ -17,7 +15,6 @@ func TestHandlerFunc_ServeKruda(t *testing.T) {
 		w.Write([]byte("ok"))
 	})
 	
-	// Create a mock request and response
 	mockReq := &mockRequest{method: "GET", path: "/test"}
 	mockResp := &mockResponseWriter{}
 	
@@ -37,83 +34,20 @@ func TestHandlerFunc_ServeKruda(t *testing.T) {
 	}
 }
 
-func TestFastHTTPIntegration(t *testing.T) {
-	var capturedMethod, capturedPath string
-	var capturedHeaders map[string]string
-	var capturedQuery map[string]string
-	
-	handler := HandlerFunc(func(w ResponseWriter, r Request) {
-		capturedMethod = r.Method()
-		capturedPath = r.Path()
-		
-		// Test provider interfaces
-		if mp, ok := r.(MultipartProvider); ok {
-			_, _ = mp.MultipartForm(1024) // Test interface
-		}
-		if cp, ok := r.(ContextProvider); ok {
-			_ = cp.Context() // Test interface
-		}
-		if ahp, ok := r.(AllHeadersProvider); ok {
-			capturedHeaders = ahp.AllHeaders()
-		}
-		if aqp, ok := r.(AllQueryProvider); ok {
-			capturedQuery = aqp.AllQuery()
-		}
-		
-		w.Header().Set("X-Test", "response")
-		w.WriteHeader(201)
-		w.Write([]byte("created"))
-	})
-	
-	// Create fasthttp context
-	ctx := &fasthttp.RequestCtx{}
-	ctx.Request.Header.SetMethod("POST")
-	ctx.Request.SetRequestURI("/api/test?param=value")
-	ctx.Request.Header.Set("Content-Type", "application/json")
-	ctx.Request.SetBody([]byte(`{"data":"test"}`))
-	
-	req := &fasthttpRequest{ctx: ctx, trustProxy: false}
-	resp := &fasthttpResponseWriter{ctx: ctx}
-	
-	handler.ServeKruda(resp, req)
-	
-	if capturedMethod != "POST" {
-		t.Errorf("expected POST, got %s", capturedMethod)
-	}
-	if capturedPath != "/api/test" {
-		t.Errorf("expected /api/test, got %s", capturedPath)
-	}
-	if capturedHeaders["Content-Type"] != "application/json" {
-		t.Errorf("expected application/json, got %s", capturedHeaders["Content-Type"])
-	}
-	if capturedQuery["param"] != "value" {
-		t.Errorf("expected value, got %s", capturedQuery["param"])
-	}
-	if ctx.Response.StatusCode() != 201 {
-		t.Errorf("expected 201, got %d", ctx.Response.StatusCode())
-	}
-	if string(ctx.Response.Body()) != "created" {
-		t.Errorf("expected created, got %s", ctx.Response.Body())
-	}
-	if string(ctx.Response.Header.Peek("X-Test")) != "response" {
-		t.Errorf("expected response, got %s", ctx.Response.Header.Peek("X-Test"))
-	}
-}
-
 // Mock implementations for testing
 type mockRequest struct {
 	method string
 	path   string
 }
 
-func (r *mockRequest) Method() string                    { return r.method }
-func (r *mockRequest) Path() string                     { return r.path }
-func (r *mockRequest) Header(key string) string         { return "" }
-func (r *mockRequest) Body() ([]byte, error)            { return nil, nil }
-func (r *mockRequest) QueryParam(key string) string     { return "" }
-func (r *mockRequest) RemoteAddr() string               { return "127.0.0.1" }
-func (r *mockRequest) Cookie(name string) string        { return "" }
-func (r *mockRequest) RawRequest() any                  { return nil }
+func (r *mockRequest) Method() string                { return r.method }
+func (r *mockRequest) Path() string                  { return r.path }
+func (r *mockRequest) Header(key string) string      { return "" }
+func (r *mockRequest) Body() ([]byte, error)         { return nil, nil }
+func (r *mockRequest) QueryParam(key string) string  { return "" }
+func (r *mockRequest) RemoteAddr() string            { return "127.0.0.1" }
+func (r *mockRequest) Cookie(name string) string     { return "" }
+func (r *mockRequest) RawRequest() any               { return nil }
 
 type mockResponseWriter struct {
 	statusCode int
